Allow arbitrary conditions to be specified via json
